Add files via upload
i have used two shallow models: MLP regerssor and SVR to forecast the electricity usage for a residential house in the Algarve. Making use of measured atmospheric air temperature and electricity consumption, sampled every 15 minutes, and daily occupation, you should forecast the house consumption for 12 hours in advance, is., 48 steps-ahead.

After using the sample parameters of each models, i have used the gridsearchCV to test with different parameters. Since the parameter tuning is taking a long time in my workstation, i couldn't imrove much. So i will work on it in future with other algorithms or techniques or models